Change library name to tiledb-segy and move package to tiledb.segy
