+++ setup.cfg
@@ -2,7 +2,7 @@
 test = pytest
 
 [coverage:run]
-omit = tilesegy/singledispatchmethod.py
+omit = tiledb/segy/singledispatchmethod.py
 
 [flake8]
 ignore = E203, E501, W503, B950
