+++ setup.py
@@ -6,19 +6,19 @@ with Path(__file__).parent.joinpath("README.md").open() as f:
     long_description = f.read()
 
 setup(
-    name="tilesegy",
-    version="0.1.0",
+    name="tiledb-segy",
+    version="0.2.0",
     description="Python library for fast access to seismic data using TileDB",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/gsakkis/tilesegy",
+    url="https://github.com/TileDB-Inc/TileDB-Segy",
     author="George Sakkis",
     author_email="george.sakkis@gmail.com",
-    packages=["tilesegy"],
+    packages=["tiledb.segy"],
     install_requires=["cached_property", "segyio>=1.9.5", "tiledb", "urlpath"],
     setup_requires=["pytest-runner"],
     tests_require=["pytest", "pytest-coverage", "pytest-xdist", "filelock"],
-    entry_points={"console_scripts": ["segy2tiledb=tilesegy.cli:main"]},
+    entry_points={"console_scripts": ["segy2tiledb=tiledb.segy.cli:main"]},
     license="MIT",
     classifiers=[
         "Development Status :: 3 - Alpha",
