+++ tiledb/segy/utils.py
